old/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
cssselect2==0.8.0
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-pgcrypto==3.0.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
Faker==37.1.0
filetype==1.2.0
fonttools==4.57.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-openai==0.3.12
langgraph==0.3.29
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.30
markdown2==2.5.3
openai==1.72.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
xxhash==3.5.0
zopfli==0.2.3.post1
zstandard==0.23.0
